remove unit summary fixed panel (incompatible with smaller screen sizes)
unit summary: •	If someone enters with has text size increased, they will not be able to access all  “select inputs” legend on the left-hand side. It does not scroll down with the scroll bar. Is there a fix for this?  Generally, I liked having this option at the top before. As an alternative, would it be possible to put radio buttons horizontal instead of vertical?